lot of struggle

+++ ner/ct_3d_projector.py
@@ -55,7 +55,7 @@ def build_conebeam_geometry(param):
     reco_space = odl.uniform_discr(min_pt=[-param.param['sx'] / 2.0, -param.param['sy'] / 2.0, -param.param['sz'] / 2.0],
                                     max_pt=[param.param['sx'] / 2.0, param.param['sy'] / 2.0, param.param['sz'] / 2.0],
                                     shape=[param.param['nx'], param.param['ny'], param.param['nz']],
-                                    dtype='float32')
+                                    dtype='float16')
 
     angle_partition = odl.uniform_partition(min_pt=param.param['start_angle'],
                                             max_pt=param.param['end_angle'],
